Fixed graph date issue - now using LONG instead of INT Problem was due to an integer overflow

+++ app/src/main/java/com/qiang/workout/Fragments/StopwatchFragment.java
@@ -148,7 +148,7 @@ public class StopwatchFragment extends Fragment
 				StopwatchTime stopwatchTime = new StopwatchTime();
 				stopwatchTime.setTime((int) timeWhenPaused / 1000);
 				stopwatchTime.setCategory(selectedCategory.getID());
-				stopwatchTime.setRecordDate((int) System.currentTimeMillis() / 1000);
+				stopwatchTime.setRecordDate(System.currentTimeMillis() / 1000);
 
 				dbHandler.addStopwatchTime(stopwatchTime);
